+++ activesupport/test/cache/cache_key_test.rb
@@ -78,6 +78,13 @@ class CacheKeyTest < ActiveSupport::TestCase
     assert_equal "foo/bar/baz", ActiveSupport::Cache.expand_cache_key(%w{foo bar baz}.to_enum)
   end
 
+  def test_set_and_get_namespace
+    cache = ActiveSupport::Cache::MemoryStore.new
+    assert_nil cache.get_namespace
+    cache.set_namespace("test")
+    assert_equal "test", cache.get_namespace
+  end
+
   private
     def with_env(kv)
       old_values = {}